Improved the Oauth handshake code Removed the login fragment (now is handled by the activity) Serialized the oauth token for persisting it through different sessions

+++ app/src/main/java/mmmarcy/github/deviantand/net/oauth/Token.java
@@ -11,7 +11,7 @@ import java.io.Serializable;
 public class Token extends GenericJson implements Serializable {
 
     @Key
-    private Integer expires_in;
+    private Double expires_in;
 
     @Key
     private String status;
